buff编辑器 - 新增被引用信息； 修复逻辑层_LateTick受到帧率影响bug； buff会对人物属性实际影响；

+++ Assets/Script/Bussiness/Logic/Action/AttributeModify/GameActionUtil_AttributeModify.cs
@@ -17,7 +17,7 @@ namespace GamePlay.Bussiness.Logic
             GameEntityBase actor,
             GameEntityBase target,
             GameAttributeType modifyType,
-            float value,
+            int value,
             GameActionValueFormat valueFormat,
             GameActionValueRefType refType
         )
